book/requirements.txt
django
gunicorn
psycopg2
djangorestframework
django-filter
